+++ src/main/java/com/trains/backend/UserOrderService.java
@@ -28,10 +28,18 @@ public class UserOrderService {
     }
 
     private void prepareStatements() {
-        SELECT_ALL_FROM_USERS_ORDERS = session.prepare("SELECT * FROM orders_per_user;");
-        INSERT_INTO_USERS_ORDERS = session.prepare("INSERT INTO orders_per_user (order_id, train_id, trip_date, user_id, car, seats_amount) VALUES (?, ?, ?, ?, ?, ?);");
-        DELETE_ALL_FROM_USERS_ORDERS = session.prepare("TRUNCATE orders_per_user;");
-        SELECT_ORDERS = session.prepare("SELECT * FROM orders_per_user WHERE train_id = ? AND trip_date = ? AND user_id = ?;");
+        if (SELECT_ALL_FROM_USERS_ORDERS == null) {
+            SELECT_ALL_FROM_USERS_ORDERS = session.prepare("SELECT * FROM orders_per_user;");
+        }
+        if (INSERT_INTO_USERS_ORDERS == null) {
+            INSERT_INTO_USERS_ORDERS = session.prepare("INSERT INTO orders_per_user (order_id, train_id, trip_date, user_id, car, seats_amount) VALUES (?, ?, ?, ?, ?, ?);");
+        }
+        if (DELETE_ALL_FROM_USERS_ORDERS == null) {
+            DELETE_ALL_FROM_USERS_ORDERS = session.prepare("TRUNCATE orders_per_user;");
+        }
+        if (SELECT_ORDERS == null) {
+            SELECT_ORDERS = session.prepare("SELECT * FROM orders_per_user WHERE train_id = ? AND trip_date = ? AND user_id = ?;");
+        }
     }
 
     public String selectAllUsersOrders() {
